pyproject.toml
[project]
name = "chat-history"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
]
